added alert

+++ app.js
@@ -28,8 +28,13 @@ function loadList(array) {
 }
 
 clear.addEventListener("click", function() {
-    localStorage.clear();
-    location.reload();
+    if (id != 0 && window.confirm("Are you sure you want to clear all your tasks?")) {
+        localStorage.clear();
+        location.reload();
+    }
+    if (id == 0) {
+        alert("Your todo list is empty");
+    }
 });
 
 const options = { weekday: "long", month: "short", day: "numeric" };
@@ -92,67 +97,44 @@ add.addEventListener("click", function() {
     input.value = "";
 });
 
-//addToDo('cofee', 1, true, false);
+
 function completeToDo(element) {
     element.classList.toggle("fa-check-circle");
     element.classList.toggle("fas");
+    element.classList.toggle("ch");
     element.classList.toggle("fa-circle");
     element.classList.toggle("far");
     console.log(element.classList);
     element.parentNode.querySelector(".text").classList.toggle(line_Through);
-
-    //let i = LIST.indexOf(element);
-    // console.log(id);
-    // console.log(i);
-
-    //const val = LIST.indexOf(element.id);
-    //console.log(val);
-    //console.log(LIST[0].name);
-
     const ind = element.attributes.id.value;
     let data = localStorage.getItem("TODO");
     const val = LIST.findIndex((x) => x.id == ind);
-    //   console.log(val);
     LIST[val].done = LIST[val].done ? false : true;
     localStorage.setItem("TODO", JSON.stringify(LIST));
 
-    //LIST = JSON.parse(data);
-    //LIST[ind].done = true;
-    //LIST[i].done = true;
-    //LIST[0].done = true;
 
-    //let data = localStorage.getItem("TODO");
-    //   console.log(data);
 }
 
 function removeToDo(element) {
     element.parentNode.parentNode.removeChild(element.parentNode);
-    //LIST[element].trash = true;
-    //let ind = LIST.indexOf(element);
     const ind = element.attributes.id.value;
-    //console.log(ind);
     const val = LIST.findIndex((x) => x.id == ind);
-    //console.log(val);
 
     LIST.splice(val, 1);
     localStorage.setItem("TODO", JSON.stringify(LIST));
-    //   console.log(LIST);
 
-    // localStorage.removeItem("TODO", LIST[element]);
 }
 
 list.addEventListener("click", function(event) {
     const element = event.target;
     const elementJob = element.attributes.job.value;
 
-    //   console.log(element);
-    //   console.log(elementJob);
+
 
     if (elementJob == "complete") {
         completeToDo(element);
     } else if (elementJob == "delete") {
         removeToDo(element);
     }
-    // localStorage.setItem("TODO", JSON.stringify(LIST));
+
 });
-//addToDo('cofee', 1, true, false);
